Update Pong.io
Our initial prototype which used the original code (the code given to us) didn't function properly for a few reasons, and one of them is that we only had one potentiometer so the paddles could only be controlled by the push buttons. This updated code allows each player to control each paddle with one potentiometer.